chore: move SongPlayer and add maps lib
